+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## UNRELEASED
+
+### Changed
+- Font textures are uploaded using power-of-two for OpenGL drivers without NPOT support
+
+
 ## 1.0.2 - 2022-01-03
 
 ### Changed